rename shared component package in home to be component.
