+++ app/src/main/java/com/amsterdam/cutetudee/presentation/screens/home/component/ImageMood.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.cutetudee.presentation.screens.home.sharedComponent
+package com.amsterdam.cutetudee.presentation.screens.home.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
