+++ app/src/main/java/com/amsterdam/cutetudee/presentation/screens/home/component/OverlayBoxContent.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.cutetudee.presentation.screens.home.sharedComponent
+package com.amsterdam.cutetudee.presentation.screens.home.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
